move TS interfaces around
